+++ Cursus/ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mflorido <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 09:43:03 by mflorido          #+#    #+#             */
-/*   Updated: 2020/02/09 18:37:30 by miguelez1        ###   ########.fr       */
+/*   Updated: 2020/02/11 19:14:42 by mflorido         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,6 +90,6 @@ t_list			*ft_lstmap(t_list *lst, void *(*f)(void *),
 
 int				ft_intlen(int n);
 void			ft_uppercase(char *str);
-char			*ft_itoh(int n);
+char			*ft_itoh(long int n);
 
 #endif
